+++ calculation.h
@@ -2,10 +2,15 @@
 #define CALCULATION_H
 
 
-class calculation
-{
-public:
-    calculation();
+class calculation { // pure virtual
+}; // average, rolling_average, etc. inherit from calculation
+
+class average : public calculation{
+
+};
+
+class rolling_average{
+
 };
 
 #endif // CALCULATION_H
