+++ loader.h
@@ -2,10 +2,15 @@
 #define LOADER_H
 
 
-class loader
-{
-public:
-    loader();
+class loader { // pure virtual
+}; // as for storage_type, text and bin loaders are children of loader
+
+class text_loader : public loader{
+
+};
+
+class bin_loader : public loader{
+
 };
 
 #endif // LOADER_H
